Change Q_GLOBAL_STATIC to generate a class rather than a single function.
This gives us exists(), isDestroyed(), destroy(), accessing as a pointer.
The whole code was moved to qglobalstatic.h in order to make qglobal.h
smaller.

The initializer support was removed (awful syntax relying on the local var
'x' in the macro, and not thread safe anyway). In the various classes using
that, the corresponding code (usually qAddPostRoutine) was moved to the ctor.