Add accessors for openssl version.
This commit adds accessors for both the openssl version number and the
version string. The intention here is to aid debugging by making it
possible for applications to know which version of the openssl library
has been found by the runtime linking code.

Since the version of openssl in use will depend on the installation of
the developer, the case cases merely display the version string and
number rather than verifying that any particular version is in use.