QString: store the hash with the QString
More than 50% of qHash are done on string for which qHash has already
been computed. Around 10% of the strings get hashed
(Source: http://pastebin.com/pbzGRJFn )
So store the hash in the string, at the price of a small overhead.

This patch also fix the alignement problem on 64bit architecture (the
string was not proberly aligned for SSE2 operations)

Currently, the previous hash algorithm is used.
It may make sens to change the hash algorithm for something faster, in
particular, the v8's hashing algorithm could be used.
Changing the hash algorithm break binary compatibility, but one could add
a flag in the reserved flag to versionize the hash.

Change-Id: Ica39f761f6ca19d89fbedfec50a4f245ad60ed04