Move the Linux futex code into a common place
I want to use futexes in QGlobalStatic too, so better not duplicate
the functionality.

This also enables the use of private futexes, which are slightly
faster. Private futexes are confined to one process (in Linux speak,
one "thread group"). Without this flag, the Linux kernel would have to
check if other processes are waiting on the same futex.