Remove the QCommandLineOptionValue class from the codebase
The existing parser class will be kept and improved. Therefore, the
Q(Core)Application becomes less bloated and more separated from the actual
command line parsing. We would need a separate class due to the convenience for
the value returnings after all.

This keeps the Q(Core)Application API simpler, and provides the option to
deprecate the parser class later in few years, if needed, due to the proper
separation. A new class or solution can be put into its place on demand.

Change-Id: I6d25b8887a0a9483f633dd5e6b90a8032cdaf5a4
Reviewed-by: David Faure <faure@kde.org>