Make the parse methods private and exclude them from the public documentation
The parsing is planned to be an internal implementation detail. The option
values can be obtained without any further need for explicit parsing. The
parsing is run internally when necessary (i.e. on demand).

The existing documentation has to be marked as internal as that should not be
part of the public documentation. On the other hand, the documentation is kept
in place and not removed for aiding the clarity and internal maintenance.

Change-Id: Ida010a693cb8d5bd74c3129b3c78805c33625f50
Reviewed-by: David Faure <faure@kde.org>