qmake: Resolve correct CC/CPLUSPLUS/LDPLUSPLUS in Xcode generator
Setting CPLUSPLUS like we used to had the consequence of Xcode trying to
run clang++++, since it has its own logic to deal with clang vs clang++.
We now set it only for the LLVM-GCC case, and make sure to set LDPLUSPLUS
so that the linker does the right thing. We also set GCC_VERSION based on
the makespecs, so that the right compiler logic is used inside Xcode,
not just the paths to the binaries.

Change-Id: Iac1625fb3c3372704ff45470a43f417478a1ef52
Reviewed-by: Morten Johan Sørvig <morten.sorvig@digia.com>