Initial mkspec and platform detection for WinRT
This establishes a foundation for a WinRT mkspec, along with notes
on which directives need to be enabled/disabled as unsupported Windows API
codepaths are removed/replaced.

The following configuration was compiled using MSVC 2012. Note that this
is a cross-compile because the tools themselves should use the normal
win32-msvc2012 mkspec:
.\configure -developer-build -xplatform winrt-msvc2012
 -no-widgets -no-accessibility -no-opengl -no-sql-sqlite
 -no-rtti -opensource -confirm-license -nomake tests -nomake demos
 -nomake examples -no-qml-debug

Change-Id: I2827c9996e7bd5d434bf8fd596b376c9f206f02a
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>