iOS: Change yet again how we determine Xcode or Makefile output
Instead of trying to magically determine if we're building Qt or
not, we base the decicion on whether or not the project depends
on QtGui and has app_bundles enabled. This prevents configure
tests and other tools from having Xcode projects, but allows
examples and demos to build out of the box.

Change-Id: I04ab474cef59d48fbc0b5bb7c726c165546072da
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@digia.com>