Map KEYCODE_BACK to Key_Back instead of Key_Close
This makes it easier to write apps/components that work on both
Android and desktop. The android back button is typically used for
the same purpose as Key_Back. Using Key_Close means having to special-
case the android version everywhere.

This is not completely ideal, since the back button also kills the app,
but I believe this is the lesser of two evils.

Change-Id: Ie39432a9564a9d28f7a11a4d4b5b37075e09d465
Related-to: QTBUG-29063
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@digia.com>