Fix debug output on Android with long signatures
The tag has a max length and is not intended to be variable
length strings. For long signatures we would thus miss the
last part of the tag, which would be the part containing the
PID, which is the part Qt Creator uses to filter out relevant
debug output. The result was that debug output from certain
functions mysteriously disappeared from the output.

Change-Id: I0221ede4e1fd2985b8c64ef0524cc2b186085afb
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>