Remove shell script from android build
The main thing the shell script did in addition to calling
configure and building was to download the OpenSSL sources
and force them into Qt. This would cause problems for people
who want to deploy their apps in countries where OpenSSL is
not available, so I've reverted that code. Instead, people
will have to do like on other platforms, and build with
dynamic resolution of the OpenSSL symbols and then bundle
the correct library with their app if they want to use the
SSL APIs.

Task-number: QTBUG-29043
Change-Id: I96ae66e53846025780823e9bd367c29967f80e03
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>
Reviewed-by: BogDan Vatra <bogdan@kde.org>