Fix compile.test on Android
By default, on Android the application template will create
a shared library, since the application on Android is actually
a Java application loading the library. This can be overridden
by using "CONFIG+=android_app" which is defined in
features/android.prf.

For compile.tests, we generally want to build regular executables,
not libraries, so we pass the configuration to the script for this
platform.

Reverts the old way of doing this, which was a hack to the
compile.test script.

Change-Id: If84cdf743f97582624b0b34ae34fd2acc8f56153
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>