Revert special case for Q_OS_ANDROID
Previously, with Qt for Android, buttons would emit clicked() if
the mouse (i.e. finger) was released outside the button. This is
not how most Android apps behave. The standard Qt desktop behaviour
is not precisely right either, but I argue that it is closer. Most
importantly, this gets rid of platform-specific code that does not
belong here.

Change-Id: I82ea26e70d7caed48ff587c5f5fcaeabfc875e3d
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@digia.com>
Reviewed-by: BogDan Vatra <bogdan@kde.org>