Remove Android-specific code
The removed code caused all widgets to take focus, so that clicking
outside an input widget would close the software keyboard. This is
not how other Android applications behave. Removing for now,
since the logic will have to be fixed anyway, and the code is in
the wrong place.

Task-number: QTBUG-29535
Change-Id: Ia1195d811078b58c4cacef14c529470ce6f5e8e4
Reviewed-by: BogDan Vatra <bogdan@kde.org>