Remove early mkspec check for Android in configure
The hack for the OpenGL ES2 compile test is no longer needed,
confirmed against a recent NDK. That means the only downside
of not determining whether we're configuring for Android early
is that we don't get an error when using the Android options in
an unrelated build. This is also the case for e.g. QNX-specific
options.

There is already a check that sets XPLATFORM_ANDROID based on
the XPLATFORM further down which matches the check for the
other platforms.

Task-number: QTBUG-29511
Change-Id: I92e19aefa66cbe283a49f6d567bcabd03ca49134
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>