Fix rotation of OpenGL-based windows on Android
The surface drawn to by Qt is not updated until swapBuffers is
called, so just sending geometry change at a random point will
not work. The new mechanics is to:

1. When Android updates its native window, send an expose event,
which at some point causes swapBuffers to be called
2. In swapBuffers, the surface is sync'ed with the native window
size (resized), so we send a geometry change event here, which,
if necessary, causes a repaint.

Change-Id: I8f0ffc6471f3bec4e5af8f1011c2dc4f39ebaf17
Reviewed-by: Samuel Rødal <samuel.rodal@digia.com>