Re-enable evdev config test
This compiles for Android, but devicediscovery, which is an
explicit requirement for the evdev code was only enabled for
linux-* platforms. So we re-enable the evdev test (which is
necessary for Android builds that use the evdev implementation)
and we add devicediscovery.pri to the Android build to make
it compile. We also disable it in the Android plugin.

Task-number: QTBUG-29688
Change-Id: Id7a2bdffd91524ddeea99b55f009b384a4fc2a57
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>