Fix configure tests for Android toolchains
The Android NDK host existence test in configure
would circumvent everything. It would show and exit
with an error even when you specified the help option,
and it would show the same error even if the problem
was that the default toolchain could not be found.

We need to move the test next to the others so that
it's tested regardless of how the architecture was specified
and so that it's not tested when help is specified, and we
need to check for the existence of the toolchain before
testing for the host architecture.

Change-Id: I5fa763ed69ea58ea025a8d98593e527ce0dd3082
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>