Move CONFIG+=android_app from configure into compile.test
The original change copied the pattern of the Mac SDK hack for
getting compile.test to work, and now that has gone away so
we should remove our hack as well. Since this is something
specific to the compile.test script and since the android_app
CONFIG will not have any effect unless android.prf is loaded,
we can just hardcode it in the script.

Change-Id: I29c8910bb36b2a1e7a26d384d09cbcaaff7207ad
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>