Fix suspend/resume on Android the right way
Suspend/resume did not work very well with the isSuspended
logic. The correct way of handling suspend/resume is to
make the window unexposed when it's suspended and then send
an expose event (which is interpreted as obscure event in
this case). We send a new expose event after the app has
been resumed again. The surface mutex is now locked recursively
since it's called from isExposed().

Change-Id: I87d3c88d457119c11df00fba80ad2128ac0678c0
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>
Reviewed-by: Gunnar Sletta <gunnar.sletta@digia.com>