Remove Necessitas icon and splash screen
We shouldn't commit these to Qt so everyone has to download
them. If we want defaults, they could perhaps be added to
Qt Creator or the SDK instead. The Android default icon is
good enough for icons though, and I'm keeping the splash.xml
so that it can easily be updated for people to add their
own splash screens.

Change-Id: Ic88158df128a6f31c207764095053eb7c5d29386
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>